splash pages working with server-fetched datasets
